requirements.txt
black
coverage
django
django-crispy-forms
django-gravatar2
django-markdown2
djangorestframework
drf_yasg
flake8
mypy
pycodestyle
python-dateutil
recommonmark==0.6.0
sly
sphinx==3.0.3
structlog
